Update monitor prefix test
Previously, this test would attempt to monitor an eval with a long
prefix, then with a too-short prefix, and finally with a
just-barely-long prefix.

Unfortunately, the last assertion sometimes does not fail correctly
(return exit status 2) because of the expectation that there is only a
single matching alloc for the given, just-barely-long, prefix. This is
not always the case as there may be another allocation that matches that
same prefix as can be seen in the reproduction below:

```
    Prefix matched multiple evaluations

ID        Priority  Type   Triggered By   Status
d3202e02  1         batch  queued-allocs  blocked
d3a0d764  1         batch  job-register   complete
```

The fix here is not to make the just-barely-long prefix assertion, as it
isn't strictly necessary.