proto: Update protos not to use GOPATH
Previously, it was required that you `go get github.com/hashicorp/nomad`
to be able to build protos, as the protoc invocation added an include
directive that pointed to `$GOPATH/src`. As Nomad now uses Go modules,
it won't necessarily be cloned to `$GOPATH`.

This change modifies the proto files and the `protoc` invocation to
handle discovering dependent protos via protoc modifier statements
specific to the protoc plugin being used.