Add fix for empty card post no matches
This fixes the bug @DingoEatingFuzz described here:
https://github.com/hashicorp/nomad/pull/9294#pullrequestreview-527748994

When any of the filter options change, an action checks whether the
active summary is excluded from the new filtered list and navigates
to another one (or the parent route) if so.

This should also happen if there is no active summary, which should
be the case when filters exclude all summaries. The getter for the
active summary needs to ignore the child controller’s model if
the child route isn’t active.