e2e: ensure Ubuntu system software is up-to-date
The cloud-init configuration will update the host's software on boot, which
can result in a race condition between that and service startup. This has
caused provisioning failures because of the Nomad unit file and systemd
version updates. Ensure that the weekly AMI build is up-to-date to reduce the
risk of this happening.