scheduler: Fix always-false sort func
I tried updating TestServiceSched_Migrate_CanaryStatus, but I wasn't
sure what the appropriate way to update it would be.

The result from s.state.DeploymentsByJobID is not deterministically
sorted, and the previous sort code ends up being a no-op as it always
returns false, but the deployment that gets selected is one that matches
the preconditions of having a matching TaskGroup that is either Promoted
or has no DesiredCanaries.

The test currently passes as it always selects the same deployment, as
only the later one matches all the preconditions.