csi: Return error when deleting node
The error that this would have returned would only have been returned if
the lookup in a map did not fail and the value that the map returned was
nil, which I think is not possible.

In this change we'll properly return the error in the
CSIPluginTypeMonolith case (which is the type given in DeleteNode()),
and also actually return the error when the given ID is not found.

This was found via errcheck.