Added new Question to FAQs
Added a popular question amongst cloud native DevOps engineers. This question seeks to help engineers in getting a proper understanding of the unique traits of Nomad - as an orchestration tool - compared to Kubernetes.

I believe the answer to this specific question lives [here](https://www.nomadproject.io/docs/nomad-vs-kubernetes). But having a summary of these key distinctions in the FAQ section (plus a URL to the page that details the comparison) would ensure a faster comprehension of these differences.