CSI: unset proto fields for volume create should be nil
The CSI spec handles empty fields and nil fields in the protobuf differently,
which may result in validation failures for creating volumes with no prior
source (and does in testing with the AWS EBS plugin). Refactor the
`CreateVolumeRequest` mapping to the protobuf in the plugin client to avoid
this bug.